backend-ai/requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow==2.10.0
pillow==10.1.0
numpy==1.23.5
python-dotenv==1.0.0
Werkzeug==3.0.1
h5py==3.8.0